# swagger_server/models/bill.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Bill(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, borrower_key: str=None, lender_key: str=None, money: int=None, info: str=None, _date: date=None, is_finaced: bool=None, is_confirmed: bool=None):  # noqa: E501
        """Bill - a model defined in Swagger

        :param borrower_key: The borrower_key of this Bill.  # noqa: E501
        :type borrower_key: str
        :param lender_key: The lender_key of this Bill.  # noqa: E501
        :type lender_key: str
        :param money: The money of this Bill.  # noqa: E501
        :type money: int
        :param info: The info of this Bill.  # noqa: E501
        :type info: str
        :param _date: The _date of this Bill.  # noqa: E501
        :type _date: date
        :param is_finaced: The is_finaced of this Bill.  # noqa: E501
        :type is_finaced: bool
        :param is_confirmed: The is_confirmed of this Bill.  # noqa: E501
        :type is_confirmed: bool
        """
        self.swagger_types = {
            'borrower_key': str,
            'lender_key': str,
            'money': int,
            'info': str,
            '_date': date,
            'is_finaced': bool,
            'is_confirmed': bool
        }

        self.attribute_map = {
            'borrower_key': 'borrowerKey',
            'lender_key': 'LenderKey',
            'money': 'money',
            'info': 'info',
            '_date': 'date',
            'is_finaced': 'isFinaced',
            'is_confirmed': 'isConfirmed'
        }

        self._borrower_key = borrower_key
        self._lender_key = lender_key
        self._money = money
        self._info = info
        self.__date = _date
        self._is_finaced = is_finaced
        self._is_confirmed = is_confirmed

    @classmethod
    def from_dict(cls, dikt) -> 'Bill':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Bill of this Bill.  # noqa: E501
        :rtype: Bill
        """
        return util.deserialize_model(dikt, cls)

    @property
    def borrower_key(self) -> str:
        """Gets the borrower_key of this Bill.


        :return: The borrower_key of this Bill.
        :rtype: str
        """
        return self._borrower_key

    @borrower_key.setter
    def borrower_key(self, borrower_key: str):
        """Sets the borrower_key of this Bill.


        :param borrower_key: The borrower_key of this Bill.
        :type borrower_key: str
        """

        self._borrower_key = borrower_key

    @property
    def lender_key(self) -> str:
        """Gets the lender_key of this Bill.


        :return: The lender_key of this Bill.
        :rtype: str
        """
        return self._lender_key

    @lender_key.setter
    def lender_key(self, lender_key: str):
        """Sets the lender_key of this Bill.


        :param lender_key: The lender_key of this Bill.
        :type lender_key: str
        """

        self._lender_key = lender_key

    @property
    def money(self) -> int:
        """Gets the money of this Bill.


        :return: The money of this Bill.
        :rtype: int
        """
        return self._money

    @money.setter
    def money(self, money: int):
        """Sets the money of this Bill.


        :param money: The money of this Bill.
        :type money: int
        """

        self._money = money

    @property
    def info(self) -> str:
        """Gets the info of this Bill.


        :return: The info of this Bill.
        :rtype: str
        """
        return self._info

    @info.setter
    def info(self, info: str):
        """Sets the info of this Bill.


        :param info: The info of this Bill.
        :type info: str
        """

        self._info = info

    @property
    def _date(self) -> date:
        """Gets the _date of this Bill.


        :return: The _date of this Bill.
        :rtype: date
        """
        return self.__date

    @_date.setter
    def _date(self, _date: date):
        """Sets the _date of this Bill.


        :param _date: The _date of this Bill.
        :type _date: date
        """

        self.__date = _date

    @property
    def is_finaced(self) -> bool:
        """Gets the is_finaced of this Bill.


        :return: The is_finaced of this Bill.
        :rtype: bool
        """
        return self._is_finaced

    @is_finaced.setter
    def is_finaced(self, is_finaced: bool):
        """Sets the is_finaced of this Bill.


        :param is_finaced: The is_finaced of this Bill.
        :type is_finaced: bool
        """

        self._is_finaced = is_finaced

    @property
    def is_confirmed(self) -> bool:
        """Gets the is_confirmed of this Bill.


        :return: The is_confirmed of this Bill.
        :rtype: bool
        """
        return self._is_confirmed

    @is_confirmed.setter
    def is_confirmed(self, is_confirmed: bool):
        """Sets the is_confirmed of this Bill.


        :param is_confirmed: The is_confirmed of this Bill.
        :type is_confirmed: bool
        """

        self._is_confirmed = is_confirmed