# swagger_server/models/bill_list.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class BillList(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self):  # noqa: E501
        """BillList - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'BillList':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BillList of this BillList.  # noqa: E501
        :rtype: BillList
        """
        return util.deserialize_model(dikt, cls)