# swagger_server/models/finace_bill.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class FinaceBill(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, money: str=None, bill_index: int=None):  # noqa: E501
        """FinaceBill - a model defined in Swagger

        :param money: The money of this FinaceBill.  # noqa: E501
        :type money: str
        :param bill_index: The bill_index of this FinaceBill.  # noqa: E501
        :type bill_index: int
        """
        self.swagger_types = {
            'money': str,
            'bill_index': int
        }

        self.attribute_map = {
            'money': 'money',
            'bill_index': 'billIndex'
        }

        self._money = money
        self._bill_index = bill_index

    @classmethod
    def from_dict(cls, dikt) -> 'FinaceBill':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The FinaceBill of this FinaceBill.  # noqa: E501
        :rtype: FinaceBill
        """
        return util.deserialize_model(dikt, cls)

    @property
    def money(self) -> str:
        """Gets the money of this FinaceBill.


        :return: The money of this FinaceBill.
        :rtype: str
        """
        return self._money

    @money.setter
    def money(self, money: str):
        """Sets the money of this FinaceBill.


        :param money: The money of this FinaceBill.
        :type money: str
        """

        self._money = money

    @property
    def bill_index(self) -> int:
        """Gets the bill_index of this FinaceBill.


        :return: The bill_index of this FinaceBill.
        :rtype: int
        """
        return self._bill_index

    @bill_index.setter
    def bill_index(self, bill_index: int):
        """Sets the bill_index of this FinaceBill.


        :param bill_index: The bill_index of this FinaceBill.
        :type bill_index: int
        """

        self._bill_index = bill_index