# swagger_server/models/transfer_bill.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class TransferBill(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, borrower_key: str=None, new_lender_key: str=None, money: int=None, info: str=None, is_confirmed: bool=None):  # noqa: E501
        """TransferBill - a model defined in Swagger

        :param borrower_key: The borrower_key of this TransferBill.  # noqa: E501
        :type borrower_key: str
        :param new_lender_key: The new_lender_key of this TransferBill.  # noqa: E501
        :type new_lender_key: str
        :param money: The money of this TransferBill.  # noqa: E501
        :type money: int
        :param info: The info of this TransferBill.  # noqa: E501
        :type info: str
        :param is_confirmed: The is_confirmed of this TransferBill.  # noqa: E501
        :type is_confirmed: bool
        """
        self.swagger_types = {
            'borrower_key': str,
            'new_lender_key': str,
            'money': int,
            'info': str,
            'is_confirmed': bool
        }

        self.attribute_map = {
            'borrower_key': 'borrowerKey',
            'new_lender_key': 'newLenderKey',
            'money': 'money',
            'info': 'info',
            'is_confirmed': 'isConfirmed'
        }

        self._borrower_key = borrower_key
        self._new_lender_key = new_lender_key
        self._money = money
        self._info = info
        self._is_confirmed = is_confirmed

    @classmethod
    def from_dict(cls, dikt) -> 'TransferBill':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TransferBill of this TransferBill.  # noqa: E501
        :rtype: TransferBill
        """
        return util.deserialize_model(dikt, cls)

    @property
    def borrower_key(self) -> str:
        """Gets the borrower_key of this TransferBill.


        :return: The borrower_key of this TransferBill.
        :rtype: str
        """
        return self._borrower_key

    @borrower_key.setter
    def borrower_key(self, borrower_key: str):
        """Sets the borrower_key of this TransferBill.


        :param borrower_key: The borrower_key of this TransferBill.
        :type borrower_key: str
        """

        self._borrower_key = borrower_key

    @property
    def new_lender_key(self) -> str:
        """Gets the new_lender_key of this TransferBill.


        :return: The new_lender_key of this TransferBill.
        :rtype: str
        """
        return self._new_lender_key

    @new_lender_key.setter
    def new_lender_key(self, new_lender_key: str):
        """Sets the new_lender_key of this TransferBill.


        :param new_lender_key: The new_lender_key of this TransferBill.
        :type new_lender_key: str
        """

        self._new_lender_key = new_lender_key

    @property
    def money(self) -> int:
        """Gets the money of this TransferBill.


        :return: The money of this TransferBill.
        :rtype: int
        """
        return self._money

    @money.setter
    def money(self, money: int):
        """Sets the money of this TransferBill.


        :param money: The money of this TransferBill.
        :type money: int
        """

        self._money = money

    @property
    def info(self) -> str:
        """Gets the info of this TransferBill.


        :return: The info of this TransferBill.
        :rtype: str
        """
        return self._info

    @info.setter
    def info(self, info: str):
        """Sets the info of this TransferBill.


        :param info: The info of this TransferBill.
        :type info: str
        """

        self._info = info

    @property
    def is_confirmed(self) -> bool:
        """Gets the is_confirmed of this TransferBill.


        :return: The is_confirmed of this TransferBill.
        :rtype: bool
        """
        return self._is_confirmed

    @is_confirmed.setter
    def is_confirmed(self, is_confirmed: bool):
        """Sets the is_confirmed of this TransferBill.


        :param is_confirmed: The is_confirmed of this TransferBill.
        :type is_confirmed: bool
        """

        self._is_confirmed = is_confirmed